+++ test/value_test.py
@@ -29,3 +29,22 @@ def test_deserialize_png_to_image():
     # Check
     assert image_value.dtype == Dtype.image
     assert image_value.shape == (224, 224, 3)
+
+def test_serialize_array_list_to_npz():
+    arrays = [randn(16000).astype("float32"), randn(32000).astype("float32")]
+    value = Value.from_object(arrays)
+    data = value.serialize("application/x-npz")
+    npz_file: NpzFile = npz_load(BytesIO(data))
+    for original, encoded in zip(arrays, npz_file.values()):
+        assert allclose(original, encoded)
+
+def test_deserialize_npz_to_array_list():
+    arrays = [randn(16000).astype("float32"), randn(32000).astype("float32")]
+    value = Value.from_object(arrays)
+    data = value.serialize("application/x-npz")
+    deserialized_value = Value.from_bytes(data, "application/x-npz")
+    result = deserialized_value.to_object()
+    assert isinstance(result, list)
+    assert len(result) == len(arrays)
+    for original, decoded in zip(arrays, result):
+        assert allclose(original, decoded)